+++ implementation/js/site/worker/counter-worker.js
@@ -14,7 +14,7 @@ var i = 0;
 function timedCount() {
     i = i + 1;
     postMessage(i);
-    setTimeout("timedCount()",500);
+    setTimeout(timedCount, 1000);
 }
 
 timedCount();